refactor: move cache persistence code into loader file (#1035)
**Context:**



**Changes In This Pull Request:**



**Test Plan:**
:
