+++ src/lib/engine/cache_loader.ts
@@ -5,7 +5,6 @@ import { TSplog } from '../utils/splog';
 import { TCachedMeta } from './cached_meta';
 import { getMetadataRefList } from './metadata_ref';
 import { parseBranchesAndMeta } from './parse_branches_and_meta';
-import { hashCacheOrKey } from './persist_cache';
 
 export const CACHE_CHECK_REF = 'refs/gt-metadata/GRAPHITE_CACHE_CHECK';
 export const CACHE_DATA_REF = 'refs/gt-metadata/GRAPHITE_CACHE_DATA';
@@ -66,4 +65,60 @@ function readPersistedCache(): Record<string, TCachedMeta> | undefined {
   } catch {
     return undefined;
   }
+}
+
+export function persistCache(
+  trunkName: string | undefined,
+  cachedBranches: Record<string, TCachedMeta>,
+  splog: TSplog
+): void {
+  splog.debug(`Persisting cache checksum to ${CACHE_CHECK_REF}...`);
+  gpExecSync(
+    {
+      command: `git update-ref ${CACHE_CHECK_REF} ${hashCacheOrKey(
+        {
+          trunkName: trunkName,
+          gitBranchNamesAndRevisions: getBranchNamesAndRevisions(),
+          metadataRefList: getMetadataRefList(),
+        },
+        true
+      )}`,
+    },
+    (err) => {
+      throw err;
+    }
+  );
+  splog.debug(`Persisting cache data to ${CACHE_DATA_REF}...`);
+  gpExecSync({
+    command: `git update-ref ${CACHE_DATA_REF} ${hashCacheOrKey(
+      cachedBranches,
+      true
+    )}`,
+  });
+  splog.debug(`Persisted cache`);
+}
+
+function hashCacheOrKey(
+  state: TCacheKey | Record<string, TCachedMeta>,
+  write?: boolean
+): string {
+  return gpExecSync(
+    {
+      command: `git hash-object ${write ? '-w' : ''} --stdin`,
+      options: {
+        input: JSON.stringify(state),
+      },
+    },
+    (err) => {
+      throw err;
+    }
+  );
+}
+
+export function clearPersistedCache(splog: TSplog): void {
+  splog.debug(`Deleting ${CACHE_CHECK_REF}...`);
+  gpExecSync({ command: `git update-ref -d ${CACHE_CHECK_REF}` });
+  splog.debug(`Deleting ${CACHE_DATA_REF}...`);
+  gpExecSync({ command: `git update-ref -d ${CACHE_DATA_REF}` });
+  splog.debug(`Cleared cache`);
 }
